add endpoint to get groups for user

+++ internal/models/group_member.go
@@ -5,7 +5,7 @@ import "time"
 type GroupMember struct {
 	CreatedAt time.Time `json:"createdAt"`
 	UpdatedAt time.Time `json:"updatedAt"`
-	UserID    uint64    `gorm:"primaryKey;autoIncrement:false" json:"userId"`
-	GroupID   uint64    `gorm:"primaryKey;autoIncrement:false" json:"groupId"`
+	UserID    uint      `gorm:"primaryKey;autoIncrement:false" json:"userId"`
+	GroupID   uint      `gorm:"primaryKey;autoIncrement:false" json:"groupId"`
 	GroupRole GroupRole `json:"groupRole"`
 }
